requirements.txt
Flask==2.3.3
pywebview==4.4.1
joblib==1.3.2
scikit-learn==1.3.0
numpy==1.24.3 